==> scripts/00-install_dependencies.r <==
#### Preamble ####
# Purpose: Install the dependencies required for this project
# Prerequisites: You must also have the R language installed
# Author: Sebastian Rodriguez
# Email: me@srod.ca
# Date: April 20, 2023
# GitHub: https://github.com/seb646/groceries-and-inflation

# Install required packages (you only need to do this once)
install.packages("lubridate")
install.packages("tidyverse")
install.packages("dplyr")
install.packages("knitr")
install.packages("janitor")
install.packages("scales")
install.packages("RColorBrewer")
install.packages("ggplot2")
install.packages("kableExtra")
install.packages("here")
install.packages("arrow")
